+++ baldwin/things/__init__.py
@@ -1,0 +1,17 @@
+"""Public package exports for Baldwin Things helpers."""
+
+from baldwin.exceptions import ThingsConfigurationError, ThingsServiceError
+
+from .client import ThingsClient
+from .models import ThingsArea, ThingsNote, ThingsProject, ThingsSnapshot, ThingsTodo
+
+__all__ = [
+    "ThingsArea",
+    "ThingsClient",
+    "ThingsConfigurationError",
+    "ThingsNote",
+    "ThingsProject",
+    "ThingsServiceError",
+    "ThingsSnapshot",
+    "ThingsTodo",
+]